refactor: decompose ASG.java into ASGEncoder + ASGDecoder delegates
Extract 1241-line ASG.java into three files:
- ASG.java (114 lines): thin façade with VERSION, ROOT_FILENAME, and
  13 one-liner delegate methods preserving the full public API
- ASGEncoder.java (506 lines): MatrixUtilities, all encode methods,
  getKey, binary/text array encoding, zip output
- ASGDecoder.java (781 lines): AbstractPropertyReference,
  PropertyReferenceToElement, all decode methods, DOM helpers,
  legacy classname conversion, binary/text array decoding

All 50 tests pass (37 decomposition + 13 model tests).
Resolves #560.
